requirements.txt
evdev==1.9.2
numpy==2.4.1
opencv-python==4.13.0.90
pillow==12.1.0
